Update CCP staging url to ccpstaging.cru.org
This DNS name points to a CloudFront distribution in front of ccpstaging.ccci.org.
The CloudFront distribution will make the public key calls less likely to fail.

Every so often, we've seen the give site receive encrypted bank accounts
that were encrypted with a stale key,
which means that the call to fetch the public key had failed
and the fallback key was used,
and the fallback key was out of date.
The public key call failure is not logged in detail,
and it's likely the browser wouldn't tell us much anyway,
so it's hard to know what the root cause has been.
Regardless, it's much more likely that a CloudFront cache will reduce failures,
because in the rare times that CCP is down the cache will probably be available.

The production url will be similarly be updated in a future commit.